+++ uberduck_ml_dev/data/get.py
@@ -85,26 +85,27 @@ def get_hubert_embeddings(audiopaths, hubert_model, output_layer = 9, hubert_pat
 
     hubert_abs_paths = []
     for audiopath in tqdm(audiopaths):
-        # NOTE (Sam): Hubert expects 16k sample rate.
-        audio0, sr = librosa.load(audiopath, sr=16000)
-        feats = torch.from_numpy(audio0)
-        feats = feats.float()
-        feats = feats.view(1, -1)
-        padding_mask = torch.BoolTensor(feats.shape).to('cpu').fill_(False)
-        inputs = {
-            "source": feats.to('cpu'),
-            "padding_mask": padding_mask,
-            "output_layer": output_layer,
-        }
-
         folder_path = str(Path(*Path(audiopath).parts[:-1]))
-        hubert_abs_path = os.path.join(folder_path , hubert_path)
-        
-        with torch.no_grad():
-            logits = hubert_model.extract_features(**inputs)
-            feats = hubert_model.final_proj(logits[0])
-            torch.save(feats[0], hubert_abs_path)
-        hubert_abs_paths.append(hubert_abs_path)
+        hubert_abs_path = os.path.join(folder_path, hubert_path)
+        # TODO (Sam): add hashing to avoid mistakenly not recomputing.
+        if not os.path.exists(hubert_abs_path):
+            # NOTE (Sam): Hubert expects 16k sample rate.
+            audio0, sr = librosa.load(audiopath, sr=16000)
+            feats = torch.from_numpy(audio0)
+            feats = feats.float()
+            feats = feats.view(1, -1)
+            padding_mask = torch.BoolTensor(feats.shape).to('cpu').fill_(False)
+            inputs = {
+                "source": feats.to('cpu'),
+                "padding_mask": padding_mask,
+                "output_layer": output_layer,
+            }
+
+            with torch.no_grad():
+                logits = hubert_model.extract_features(**inputs)
+                feats = hubert_model.final_proj(logits[0])
+                torch.save(feats[0], hubert_abs_path)
+            hubert_abs_paths.append(hubert_abs_path)
 
     return hubert_abs_paths
